+++ crates/burn-tensor/Cargo.toml
@@ -41,6 +41,7 @@ rand = { workspace = true }
 rand_distr = { workspace = true }                    # use instead of statrs because it supports no_std
 bytemuck = { workspace = true }
 colored = { workspace = true, optional = true }
+spin = { workspace = true }
 
 # The same implementation of HashMap in std but with no_std support (only needs alloc crate)
 hashbrown = { workspace = true } # no_std compatible
