+++ crates/burn-tensor/src/router/bridge/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod byte;
+
+pub use base::*;
+pub use byte::*;
