+++ crates/burn-tensor/src/router/channel/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod direct;
+
+pub use base::*;
+pub use direct::*;
