+++ crates/burn-tensor/src/router/client/mod.rs
@@ -1,0 +1,3 @@
+mod base;
+
+pub use base::*;
