+++ crates/burn-tensor/src/router/mod.rs
@@ -1,8 +1,14 @@
 mod backend;
+mod bridge;
+mod channel;
 mod client;
 mod ops;
+mod runner;
 mod tensor;
 
 pub use backend::*;
+pub use bridge::*;
+pub use channel::*;
 pub use client::*;
+pub use runner::*;
 pub use tensor::*;
