+++ crates/burn-tensor/src/router/ops/mod.rs
