+++ crates/burn-tensor/src/router/ops/op_activation.rs
@@ -1,0 +1,4 @@
+use crate::ops::ActivationOps;
+use crate::router::{BackendRouter, RunnerChannel};
+
+impl<R: RunnerChannel> ActivationOps<Self> for BackendRouter<R> {}
