+++ crates/burn-tensor/src/router/tensor.rs
