--- crates/burn-tensor/src/runner/client/mod.rs
@@ -1,1 +1,0 @@
-pub mod in_memory;
