--- crates/burn-tensor/src/runner/ops/op_activation.rs
@@ -1,4 +1,0 @@
-use crate::ops::ActivationOps;
-use crate::runner::{BackendRouter, RunnerChannel};
-
-impl<C: RunnerChannel> ActivationOps<Self> for BackendRouter<C> {}
